Move the logic to execute a navigation to a separate module
