+++ src/utilities/index.ts
@@ -3,6 +3,7 @@ export { getTemplateWithPartials } from './modules/partials';
 export { getCSSVariables } from './modules/css';
 export { randomId, parseWidth } from './modules/misc';
 export { getRestApis } from './modules/restApi';
+export * from './modules/navigate';
 export * from './modules/dialogs';
 export * from './modules/predicates';
 export { getFormatDateMethods } from './modules/format-date';
